requirements.txt
faker
python-dotenv
tqdm
